move core api features into a separate class
